another set of refactoring
