+++ planetmint/abci/block.py
@@ -1,0 +1,3 @@
+from collections import namedtuple
+
+Block = namedtuple("Block", ("app_hash", "height", "transactions"))
