+++ tests/elections/test_election.py
@@ -1,7 +1,7 @@
 import pytest
 
 from tests.utils import generate_election, generate_validators
-from planetmint.lib import Block
+from planetmint.abci.block import Block
 from transactions.types.elections.election import Election
 from transactions.types.elections.chain_migration_election import ChainMigrationElection
 from transactions.types.elections.validator_election import ValidatorElection
